api/db.py
# api/db.py
from sqlalchemy import create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

DB_USER = os.getenv("DB_USER", "postgres")
DB_PASS = os.getenv("DB_PASS", "admin")
DB_HOST = os.getenv("DB_HOST", "localhost")
DB_PORT = os.getenv("DB_PORT", "5432")
DB_NAME = os.getenv("DB_NAME", "TransactAI")

# Support direct DATABASE_URL injection (common in production like Render/Heroku)
DATABASE_URL = os.getenv("DATABASE_URL")

if not DATABASE_URL:
    # Use SQLite as a zero-config fallback for local development if PG is not explicitly required
    # This fixes "Network Error" and backend crashes when local PostgreSQL is not running
    if DB_HOST == "localhost" and not os.getenv("FORCE_POSTGRES"):
        print("WARNING: Local PostgreSQL not detected or configured. Falling back to SQLite for local development.")
        DATABASE_URL = "sqlite:///./transactai.db"
    else:
        DATABASE_URL = f"postgresql://{DB_USER}:{DB_PASS}@{DB_HOST}:{DB_PORT}/{DB_NAME}"

# Engine Initialization
# SQLite needs different connect_args for multi-threading
connect_args = {"check_same_thread": False} if DATABASE_URL.startswith("sqlite") else {}

engine = create_engine(
    DATABASE_URL, 
    echo=True,
    connect_args=connect_args
)

SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

Base = declarative_base()

# Dependency for routes
def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()
