requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
scikit-learn==1.2.2
pandas==2.0.3
numpy==1.25.2
joblib==1.2.0
nltk==3.8.1
python-multipart==0.0.6
matplotlib==3.8.1
seaborn==0.12.2
typing-extensions==4.8.0
imbalanced-learn==0.11.0
torch==2.3.0
transformers==4.41.2
accelerate==0.31.0
sentence-transformers==2.7.0
rapidfuzz==3.1.1
psycopg2-binary
sqlalchemy
requests
python-dotenv
regex